Fixed another chain removal issue again (probably)
I screwed up when making the Chain class and it's a ton of work to fix
it off of remote desktop on my phone right now. Instead of making
methods in First and Second, which remove the objects, link them, or get
other objects linked to them, I just managed the Chain by manipulatin
Firsts' and Seconds' references to their successors.

So now it's a pain in the ass to handle all the possible edits (in an
iterator, for example) that could be applied to the Chain
simultaneously. This class is very badly made.

I don't even know if it still has issues or not, yet.